+++ src/main/java/net/sybyline/scarlet/ScarletPendingModActions.java
@@ -21,15 +21,26 @@ public class ScarletPendingModActions
     {
         this.pendingModActionsFile = pendingModActionsFile;
         this.pendingModActions = new ConcurrentHashMap<>();
+        this.pendingBanInfo = new ConcurrentHashMap<>();
         this.load();
     }
 
     final File pendingModActionsFile;
     final Map<String, String> pendingModActions;
+    final Map<String, BanInfo> pendingBanInfo;
 
     public static class DataSpec
     {
         public Map<String, String> pendingModActions;
+        public Map<String, BanInfo> pendingBanInfo;
+    }
+    public static class BanInfo
+    {
+        public BanInfo()
+        {
+        }
+        public String description;
+        public String[] tags;
     }
 
     public boolean hasPending(GroupAuditType type, String targetUserId)
@@ -62,6 +73,42 @@ public class ScarletPendingModActions
         return prevActorUserId;
     }
 
+    public boolean addBanInfo(String targetUserId)
+    {
+        if (this.pendingBanInfo.putIfAbsent(targetUserId, new BanInfo()) != null)
+            return false;
+        this.save();
+        return true;
+    }
+
+    public boolean setBanInfoTags(String targetUserId, String[] tags)
+    {
+        BanInfo info = this.pendingBanInfo.get(targetUserId);
+        if (info == null)
+            return false;
+        info.tags = tags;
+        this.save();
+        return true;
+    }
+
+    public boolean setBanInfoDescription(String targetUserId, String description)
+    {
+        BanInfo info = this.pendingBanInfo.get(targetUserId);
+        if (info == null)
+            return false;
+        info.description = description;
+        this.save();
+        return true;
+    }
+
+    public BanInfo pollBanInfo(String targetUserId)
+    {
+        BanInfo info = this.pendingBanInfo.remove(targetUserId);
+        if (info != null)
+            this.save();
+        return info;
+    }
+
     public boolean load()
     {
         if (!this.pendingModActionsFile.isFile())
@@ -84,6 +131,11 @@ public class ScarletPendingModActions
         {
             this.pendingModActions.putAll(spec.pendingModActions);
         }
+        this.pendingBanInfo.clear();
+        if (spec != null && spec.pendingBanInfo != null && !spec.pendingBanInfo.isEmpty())
+        {
+            this.pendingBanInfo.putAll(spec.pendingBanInfo);
+        }
         return true;
     }
 
@@ -91,6 +143,7 @@ public class ScarletPendingModActions
     {
         DataSpec spec = new DataSpec();
         spec.pendingModActions = new HashMap<>(this.pendingModActions);
+        spec.pendingBanInfo = new HashMap<>(this.pendingBanInfo);
         try (Writer w = MiscUtils.writer(this.pendingModActionsFile))
         {
             Scarlet.GSON_PRETTY.toJson(spec, DataSpec.class, w);